Alerting: use Mimir logo
